requirements.txt
boto3==1.16.10
crispy-tailwind==0.3.0
django_heroku
Django==3.2.6
django-crispy-forms==1.9.2
django-storages==1.10.1
gunicorn==20.1.0
Pillow==8.3.1
python-decouple==3.4
requests==2.26.0
pandas
pyvis
pymongo
dnspython
openpyxl
plotly==5.4.0
